Move utility types and functions currently defined in composable-controller to base-controller
