+++ packages/base-controller/src/BaseControllerV1.ts
@@ -1,3 +1,34 @@
+import type { PublicInterface } from '@metamask/utils';
+
+import type { ControllerInstance } from './BaseControllerV2';
+
+/**
+ * Determines if the given controller is an instance of `BaseControllerV1`
+ *
+ * @param controller - Controller instance to check
+ * @returns True if the controller is an instance of `BaseControllerV1`
+ */
+export function isBaseControllerV1(
+  controller: ControllerInstance,
+): controller is BaseControllerV1Instance {
+  return (
+    'name' in controller &&
+    typeof controller.name === 'string' &&
+    'config' in controller &&
+    typeof controller.config === 'object' &&
+    'defaultConfig' in controller &&
+    typeof controller.defaultConfig === 'object' &&
+    'state' in controller &&
+    typeof controller.state === 'object' &&
+    'defaultState' in controller &&
+    typeof controller.defaultState === 'object' &&
+    'disabled' in controller &&
+    typeof controller.disabled === 'boolean' &&
+    'subscribe' in controller &&
+    typeof controller.subscribe === 'function'
+  );
+}
+
 /**
  * State change callbacks
  */
@@ -30,6 +61,26 @@ export interface BaseConfig {
 export interface BaseState {
   name?: string;
 }
+
+/**
+ * The narrowest supertype for `BaseControllerV1` config objects.
+ * This type can be assigned to any `BaseControllerV1` config object.
+ */
+export type ConfigConstraint = BaseConfig & object;
+
+/**
+ * The narrowest supertype for `BaseControllerV1` state objects.
+ * This type can be assigned to any `BaseControllerV1` state object.
+ */
+export type StateConstraint = BaseState & object;
+
+/**
+ * The widest subtype of all controller instances that extend from `BaseControllerV1`.
+ * Any `BaseControllerV1` instance can be assigned to this type.
+ */
+export type BaseControllerV1Instance = PublicInterface<
+  BaseControllerV1<ConfigConstraint, StateConstraint>
+>;
 
 /**
  * @deprecated This class has been renamed to BaseControllerV1 and is no longer recommended for use for controllers. Please use BaseController (formerly BaseControllerV2) instead.
