+++ packages/base-controller/src/index.ts
@@ -1,11 +1,24 @@
-export type { BaseConfig, BaseState, Listener } from './BaseControllerV1';
-export { BaseControllerV1 } from './BaseControllerV1';
 export type {
+  BaseConfig,
+  BaseControllerV1Instance,
+  BaseState,
+  ConfigConstraint as ConfigConstraintV1,
+  Listener,
+  StateConstraint as StateConstraintV1,
+} from './BaseControllerV1';
+export { BaseControllerV1, isBaseControllerV1 } from './BaseControllerV1';
+export type {
+  BaseControllerInstance,
+  ControllerInstance,
   Listener as ListenerV2,
   StateConstraint,
+  LegacyControllerStateConstraint,
   StateDeriver,
+  StateDeriverConstraint,
   StateMetadata,
+  StateMetadataConstraint,
   StatePropertyMetadata,
+  StatePropertyMetadataConstraint,
   ControllerGetStateAction,
   ControllerStateChangeEvent,
 } from './BaseControllerV2';
@@ -13,6 +26,8 @@ export {
   BaseController,
   getAnonymizedState,
   getPersistentState,
+  isBaseController,
 } from './BaseControllerV2';
 export * from './ControllerMessenger';
-export * from './RestrictedControllerMessenger';
+export type { RestrictedControllerMessengerConstraint } from './RestrictedControllerMessenger';
+export { RestrictedControllerMessenger } from './RestrictedControllerMessenger';
